requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
firebase-admin>=6.0.0
flask-compress==1.23
python-dotenv==1.0.0

gunicorn>=21.0.0
